requirements_LOCK.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==5.1.0
certifi==2019.11.28
chardet==3.0.4
dbus-python==1.2.16
distro==1.4.0
ffmpeg-python==0.2.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.6.0
idna==2.8
importlib-metadata==4.11.4
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyGObject==3.36.0
pymongo==4.2.0
pyparsing==3.0.9
python-apt==2.0.0+ubuntu0.20.4.7
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
scipy==1.9.1
six==1.14.0
ssh-import-id==5.10
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
typing_extensions==4.2.0
urllib3==1.25.8
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
